move Kafka application into kafka folder
